--- back/reserva-tech/src/components/Colaborador.css
@@ -1,155 +1,0 @@
-* {
-  box-sizing: border-box;
-  font-family: Arial, Helvetica, sans-serif;
-}
-
-body {
-  margin: 0;
-  background-color: #ffffff;
-}
-
-.container {
-  max-width: 1200px;
-  margin: 50px auto;
-  padding: 40px;
-  background-color: #dedede;
-  border-radius: 8px;
-  border: 2px solid #8b78ff;
-}
-
-.header {
-  display: flex;
-  align-items: center;
-  gap: 20px;
-  border-bottom: 2px solid #5f5f5f;
-  padding-bottom: 20px;
-  margin-bottom: 40px;
-}
-
-.header h1 {
-  flex: 1;
-  text-align: center;
-  margin: 0;
-  font-size: 28px;
-  font-weight: 500;
-  color: #000000;
-}
-
-.back-button {
-  background: none;
-  border: none;
-  font-size: 32px;
-  cursor: pointer;
-  color: #0044ff;
-  padding: 0 10px;
-}
-
-.back-button:hover {
-  opacity: 0.7;
-}
-
-.form-container {
-  display: flex;
-  gap: 50px;
-}
-
-.column {
-  flex: 1;
-}
-
-.divider {
-  width: 2px;
-  background-color: #5f5f5f;
-}
-
-.field {
-  display: flex;
-  flex-direction: column;
-  margin-bottom: 35px;
-}
-
-.field label {
-  margin-bottom: 8px;
-  font-size: 18px;
-  color: #000000;
-  font-weight: 500;
-}
-
-.field input {
-  height: 60px;
-  border-radius: 30px;
-  border: none;
-  padding: 0 25px;
-  font-size: 18px;
-  outline: none;
-  background-color: #ffffff;
-  box-shadow: 0 4px 6px rgba(116, 116, 116, 0.25);
-  color: #000000; 
-}
-
-.field input::placeholder {
-  color: #666666;
-  opacity: 0.8;
-}
-
-.field input:focus {
-  box-shadow: 0 4px 8px rgba(124, 107, 255, 0.3);
-}
-
-.field input[type="date"] {
-  color: #000000; 
-}
-
-.field input[type="date"]::-webkit-calendar-picker-indicator {
-  filter: invert(0); 
-  cursor: pointer;
-}
-
-.submit-button {
-  display: block;
-  margin: 50px auto 0;
-  padding: 18px 60px;
-  background-color: #7c6bff;
-  color: #ffffff;
-  border: none;
-  border-radius: 12px;
-  font-size: 22px;
-  font-weight: 500;
-  cursor: pointer;
-  transition: background-color 0.2s;
-}
-
-.submit-button:hover {
-  background-color: #6a5ae0;
-}
-
-@media (max-width: 768px) {
-  .container {
-    margin: 20px;
-    padding: 20px;
-  }
-  
-  .form-container {
-    flex-direction: column;
-    gap: 20px;
-  }
-  
-  .divider {
-    width: 100%;
-    height: 2px;
-  }
-  
-  .header h1 {
-    font-size: 24px;
-  }
-  
-  .field input {
-    height: 50px;
-    font-size: 16px;
-  }
-  
-  .submit-button {
-    padding: 15px 40px;
-    font-size: 18px;
-  }
-}
